+++ README.md
@@ -44,7 +44,8 @@ _The account number must have a corresponding key in [`accounts.js`](./accounts.
 1. Create `.env` file with print proxy wide open quad-word.
 1. `touch .env && echo 'OPEN_QUAD_WORD=<wide-open-quad-word>' >> .env`
 1. `npm install` & `npm start`
-1. `node scripts/seed-firestore-emulator.js`
+1. `node scripts/seed-test-data.js`
+1. `node scripts/pull-accounts-from-staging.js`
 
 ## Testing
